test(strkern): two standalone harnesses become one doctest target beside verify_csr
The owner's form for a C++ unit gate is test/verify_csr.cpp / test/verify_pagerank.cpp:
DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN, TEST_CASE/CHECK/REQUIRE, a CMake target under -DRIPWIRE_TESTS=ON,
doctest vendored at third_party/deps/doctest. test/strkern_harness.cpp and test/emitescape_harness.cpp
were the other form — hand-rolled `checkf` printers with their own main and their own PASS counting —
and they tested the SAME header from two sides: the escapers are strkern::findByteset's only shipped
callers, so a set bug and a scan bug are indistinguishable from a diff. One file, one target.

  test/verify_strkern.cpp    19 TEST_CASEs, 19 assertions, one per kernel and per escaper arm.
  CMake ripwire_test_strkern beside ripwire_test_csr: doctest::doctest, the same include paths, added
                             to RIPWIRE_RUNTIME_COMPILE/LINK and RIPWIRE_OWNED_CXX_TARGETS so
                             -DRIPWIRE_ASAN=ON carries the complete G1 stack onto it, `add_test(NAME
                             ripwire.strkern)` so ctest runs it.

ARM COUNT, BEFORE AND AFTER, printed by both gates on every run so a lost arm is arithmetic:
14 (strkern_harness) + 4 (emitescape_harness) = 18 -> 19. The extra one is the compiled-path
assertion; nothing else was added or dropped. Preserved exactly: both fixed-seed corpora (100k random
buffers over four alphabets x lengths 0..300; 222,682 adversarial escaper inputs), the
all-256-bytes-at-every-offset-and-length sweep, the every-byte-of-src/-and-docs/ real-text arms, the
VERBATIM pre-2026-09-10 tokenizer walkers and the VERBATIM pre-rewrite escapers as oracles, and BOTH
mutation controls (-DSTRKERN_MUTATE=1, -DEMITESCAPE_MUTATE_BYTESET=1).

Each corpus is walked ONCE in a memoised builder the TEST_CASEs read, so one-assertion-per-test-case
costs no extra pass over 100k buffers. The old harnesses' bundled arms became per-kernel probe
functions (probeClassMasks / probeFold / probeFoldedEquals / probeFindByte / probeFind3 /
probeFindByteset, and probeFile* for the real-text side), which is where splitting was worth doing.
The random draws stay inside the walk in the same ORDER, so every arm sees the same corpus it did.

THE GATES DRIVE THE TARGET.
  test/strkerncheck.sh   (1) cmake -DRIPWIRE_TESTS=ON -DRIPWIRE_ASAN=ON in a scratch dir +
                             --target ripwire_test_strkern, run whole: the G1 arm for the entire TU,
                             kernels and escapers, with -fno-sanitize-recover=all. (2) direct $CXX
                             -DSTRKERN_MUTATE=1, must fail: 11 of 19 red (was 8, then 9 once the tail
                             fix added the words mutation; 11 now that the escapers route through
                             findByteset, so a kernel mutation reaches them). (3) direct $CXX
                             -arch x86_64 -march=x86-64-v3 under Rosetta 2, 19 assertions green.
  test/emitescapecheck.sh (A) the same CMake target, `-tc=escape:*`, 4 test cases / 4 assertions.
                          (B) direct $CXX -DEMITESCAPE_MUTATE_BYTESET=1: 168,423 of 222,682 inputs
                              must disagree — same number the standalone harness reported.
                          (C) the every-byte fixture through --for/--expand/--json, unchanged.
  Arms 2, 3 and B are DIRECT COMPILES on purpose, and the scripts say why: CMake cannot express a
  second architecture for one target in this tree, and a second full configure to pass one -D would
  cost a configure to say nothing extra. emitescapecheck builds WITHOUT sanitizers because
  strkerncheck already runs every test case in this TU under the complete stack; a second sanitized
  copy would re-prove that at the price of another build.

test/binoverridecheck.sh's EXEMPT reason for strkerncheck is updated (it drives a CMake target now,
and still never invokes build/ripwire). No new gate file, so test/regression.sh and the published gate
count are unchanged — gatecountcheck and manifestcheck green.

COST, stated because it went up: strkerncheck 17 s -> 80 s wall / 48 s CPU, emitescapecheck ~15 s ->
27 s wall / 21 s CPU, measured at box load 27. The TU is bigger (it pulls src/serialize.h) and is
compiled three times. No pargates.py GATE_BUDGET_SEC entry is added: 48 s of CPU against a 300 s
default (x4 on CI's --budget-scale) is 6x headroom, and an entry for it would have been this commit's
only gating --quality-delta row. If CI ever shows the rc=124-at-exactly-the-cap signature for either
gate, the entry is the remedy and these are the numbers it should carry.

--quality-delta: gating=0, preexisting-worse=0, 25 new-symbol rows, nothing acked. One of them was a
real clone the tool was right about (classMasksSweep vs probeClassMasks shared a 207-token compare)
and is fixed by funnelling both through compareClassMasksWindow. The rest are read and kept, with the
reasons at the top of the file: the frozen *Ref oracles must not be restructured to flatter a metric,
and the repeated 36-43 token TEST_CASE body IS one-assertion-per-test-case.

--test-gate names exactly the three changed gates; untested=0; all three green. Byte-identical 21/21
(3 corpora x 7 surfaces), determinism and xmllint clean — no shipped code moved in this commit.
